Atualização do arquivo da dag de produção bio geral

+++ db/schemas/trusted/ddl_producao_biodiesel_m3_geral.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS td_ext_anp.producao_biodiesel_m3_geral (
     unidade_federacao STRING OPTIONS(description="Estado do Brasil"),
     produtor STRING OPTIONS(description="Nome do produtor"),
     produto STRING OPTIONS(description="Nome do produto"),
-    producao FLOAT64 OPTIONS(description="Quantidade do produto em m³"),
+    producao NUMERIC OPTIONS(description="Quantidade do produto em m³"),
     data_criacao TIMESTAMP OPTIONS(description="Data de inserção dos dados na camada raw")
     data_ingestao_td TIMESTAMP DEFAULT CURRENT_TIMESTAMP() OPTIONS(description="Data de inserção dos dados na camada trusted")
 ) PARTITION BY DATE(data_ingestao_td);
